fix(gcp): suppress private key diff to avoid updates
When you run terraform apply after creating a gcp resource, we were
trying to update the `private_key` when it is a sensitive element and
the Lacework server, for security reasons, doesn't return that secret,
which means that the resource was always different.

We are adding a suppression to avoid updating the resource.

Closes https://github.com/terraform-providers/terraform-provider-lacework/issues/4

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>