fix(azure): suppress client_secret diff to avoid updates
When you run terraform apply after creating a azure resource, we were
trying to update the client_secret when it is a sensitive element and
the Lacework server, for security reasons, doesn't return that secret,
which means that the resource was always different.

We are adding a suppression to avoid updating the resource.

Signed-off-by: Salim Afiune Maya <afiune@lacework.net>